+++ CHANGELOG.md
@@ -74,6 +74,9 @@
   non-null value passed into the initializer.
 * **\[gwt-output-qa\]** Add assertion to `arez.gwt.qa.ArezBuildAsserts` that ensures that the method
   `arez.Observable.preReportChanged()` is not present in production code.
+* **\[gwt-output-qa\]** Add assertion `arez.gwt.qa.ArezBuildAsserts.assertEquals(...)` that can be used to
+  ensure generated classes do or do not have equals defined. Add assertions into standard asserts that ensure
+  that no repositories have an `equals(...)` method defined.
 
 ### [v0.104](https://github.com/arez/arez/tree/v0.104) (2018-08-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.103...v0.104)